+++ docs/changelog.rst
@@ -264,3 +264,6 @@ Internals
   for example when evaluating the :meth:`~capytaine.bem.solver.compute_free_surface_elevation` along the waterline.
 
 * New function implemented :func:`~capytaine.meshes.water_line_integral` to compute the integral of a function along the water line of a mesh.
+
+* "Full double-layer matrices" are now stored as array of shape (3, n, m)
+  (or 3 matrices of shape (n, m)) instead of (n, m, 3) (:pull:`869`).